+++ submodules/TelegramCore/Sources/TelegramEngine/Peers/TelegramEnginePeers.swift
@@ -51,7 +51,7 @@ public extension TelegramEngine {
             }
         }
 
-        public func findChannelById(channelId: Int32) -> Signal<Peer?, NoError> {
+        public func findChannelById(channelId: Int64) -> Signal<Peer?, NoError> {
             return _internal_findChannelById(postbox: self.account.postbox, network: self.account.network, channelId: channelId)
         }
